Redis bucketer now sets timeout timers on startup
Added support for redis bucketer to find all known buckets on startup
and set timeout timers for every bucket. This ensures that even if the
app is restarted the redis bucketer will still find all buckets and have
timers. This is important since otherwise you may lose buckets if the
app is restarted and then nothing ever gets added to the bucket again.